[repacker] Begin splitting up the repacker implementation into several files.
