+++ src/test-repacker.cc
@@ -28,6 +28,7 @@
 
 #include "hb-repacker.hh"
 #include "hb-open-type.hh"
+#include "graph/serialize.hh"
 
 static void start_object(const char* tag,
                          unsigned len,
@@ -931,7 +932,7 @@ test_serialize ()
   hb_bytes_t expected = c1.copy_bytes ();
 
   graph_t graph (c1.object_graph ());
-  hb_blob_t* out = graph.serialize ();
+  hb_blob_t* out = graph::serialize (graph);
   free (buffer_1);
 
   hb_bytes_t actual = out->as_bytes ();
